move Vivarium and add get_timeseries
